# Conflicts:
#	backend/de.metas.business/src/test/java/de/metas/project/workorder/calendar/WOProjectCalendarServiceTest.java
#	backend/de.metas.business/src/test/java/de/metas/project/workorder/calendar/WOProjectSimulationPlanEditorTest.java

+++ backend/de.metas.ui.web.base/src/main/java/de/metas/ui/web/window/model/DocumentCollection.java
@@ -30,6 +30,7 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
 import de.metas.document.references.zoom_into.RecordWindowFinder;
 import de.metas.i18n.AdMessageKey;
+import de.metas.i18n.BooleanWithReason;
 import de.metas.letters.model.MADBoilerPlate;
 import de.metas.letters.model.MADBoilerPlate.BoilerPlateContext;
 import de.metas.letters.model.MADBoilerPlate.SourceDocument;
@@ -141,7 +142,8 @@ public class DocumentCollection
 	/**
 	 * Delegates to the {@link DocumentDescriptorFactory#isWindowIdSupported(WindowId)} of this instance's {@code documentDescriptorFactory}.
 	 */
-	public boolean isWindowIdSupported(@Nullable final WindowId windowId)
+	public boolean isWindowIdSupported(
+			@Nullable final WindowId windowId)
 	{
 		return documentDescriptorFactory.isWindowIdSupported(windowId);
 	}
@@ -175,12 +177,15 @@ public class DocumentCollection
 		return windowIds != null && !windowIds.isEmpty() ? ImmutableSet.copyOf(windowIds) : ImmutableSet.of();
 	}
 
-	public Document getDocumentReadonly(@NonNull final DocumentPath documentPath)
+	public Document getDocumentReadonly(
+			@NonNull final DocumentPath documentPath)
 	{
 		return forDocumentReadonly(documentPath, Function.identity());
 	}
 
-	public <R> R forDocumentReadonly(@NonNull final DocumentPath documentPath, @NonNull final Function<Document, R> documentProcessor)
+	public <R> R forDocumentReadonly(
+			@NonNull final DocumentPath documentPath,
+			@NonNull final Function<Document, R> documentProcessor)
 	{
 		final DocumentPath rootDocumentPath = documentPath.getRootDocumentPath();
 
@@ -204,7 +209,8 @@ public class DocumentCollection
 		});
 	}
 
-	private Document getOrLoadDocument(@NonNull final DocumentKey documentKey)
+	private Document getOrLoadDocument(
+			@NonNull final DocumentKey documentKey)
 	{
 		try
 		{
@@ -223,7 +229,8 @@ public class DocumentCollection
 		}
 	}
 
-	public <R> R forRootDocumentReadonly(@NonNull final DocumentPath documentPath, final Function<Document, R> rootDocumentProcessor)
+	public <R> R forRootDocumentReadonly(
+			@NonNull final DocumentPath documentPath, final Function<Document, R> rootDocumentProcessor)
 	{
 		final DocumentKey rootDocumentKey = DocumentKey.ofRootDocumentPath(documentPath.getRootDocumentPath());
 
@@ -365,7 +372,8 @@ public class DocumentCollection
 	/**
 	 * Retrieves document from repository
 	 */
-	private Document retrieveRootDocumentFromRepository(final DocumentKey documentKey)
+	private Document retrieveRootDocumentFromRepository(
+			@NonNull final DocumentKey documentKey)
 	{
 		final DocumentEntityDescriptor entityDescriptor = getDocumentEntityDescriptor(documentKey.getWindowId());
 
@@ -402,14 +410,14 @@ public class DocumentCollection
 			long countDocumentsWithChanges = 0;
 			final List<DocumentKey> documentKeysToInvalidate = new ArrayList<>();
 			for (final Map.Entry<DocumentKey, Document> entry : rootDocuments.asMap().entrySet())
-			{
+				{
 				final Document document = entry.getValue();
 				if (document.hasChangesRecursivelly())
-				{
-					countDocumentsWithChanges++;
-				}
-				else
-				{
+					{
+						countDocumentsWithChanges++;
+					}
+					else
+					{
 					documentKeysToInvalidate.add(entry.getKey());
 				}
 			}
@@ -426,7 +434,8 @@ public class DocumentCollection
 		return result;
 	}
 
-	private void commitRootDocument(@NonNull final Document rootDocument)
+	private void commitRootDocument(
+			@NonNull final Document rootDocument)
 	{
 		Preconditions.checkState(rootDocument.isRootDocument(), "{} is not a root document", rootDocument);
 
@@ -449,17 +458,12 @@ public class DocumentCollection
 
 		//
 		// Make sure all events were collected for the case when we just created the new document
-		// FIXME: this is a workaround and in case we find out all events were collected, we just need to remove this.
+		//
+		// IMPORTANT: This case happens when the document is not yet saved, but it's not the first time we are patching it.
+		// e.g. we have a document with multiple mandatory fields, user is filling them one by one, after each change a PATCH is sent
 		if (wasNew)
 		{
-			logger.debug("Checking if we collected all events for the new document");
-			final Set<String> collectedFieldNames = rootDocument.getChangesCollector().collectFrom(rootDocument, () -> "new document, initially missed");
-			if (!collectedFieldNames.isEmpty())
-			{
-				logger.warn("We would expect all events to be auto-magically collected but it seems that not all of them were collected!"
-						+ "\n Missed (but collected now) field names were: {}" //
-						+ "\n Document path: {}", collectedFieldNames, rootDocument.getDocumentPath());
-			}
+			rootDocument.getChangesCollector().collectFrom(rootDocument, () -> "new document, initially missed");
 		}
 
 	}
@@ -479,8 +483,16 @@ public class DocumentCollection
 		}
 
 		forRootDocumentWritable(rootDocumentPath, changesCollector, rootDocument -> {
+
 			if (documentPath.isRootDocument())
 			{
+				final BooleanWithReason isDeleteForbidden = rootDocument.isDeleteForbidden();
+				if (isDeleteForbidden.isTrue())
+				{
+					throw new AdempiereException(isDeleteForbidden.getReason())
+							.markAsUserValidationError();
+				}
+
 				if (!rootDocument.isNew())
 				{
 					rootDocument.deleteFromRepository();
@@ -501,7 +513,8 @@ public class DocumentCollection
 		});
 	}
 
-	private void assertDeleteDocumentAllowed(final DocumentEntityDescriptor entityDescriptor)
+	private void assertDeleteDocumentAllowed(
+			@NonNull final DocumentEntityDescriptor entityDescriptor)
 	{
 		final Evaluatee evalCtx = Evaluatees.mapBuilder()
 				.put(WindowConstants.FIELDNAME_Processed, false)
@@ -530,7 +543,8 @@ public class DocumentCollection
 		return documentDescriptorFactory.getTableRecordReference(documentPath);
 	}
 
-	public WindowId getWindowId(@NonNull final DocumentZoomIntoInfo zoomIntoInfo)
+	public WindowId getWindowId(
+			@NonNull final DocumentZoomIntoInfo zoomIntoInfo)
 	{
 		if (zoomIntoInfo.getWindowId() != null)
 		{
@@ -570,7 +584,8 @@ public class DocumentCollection
 	/**
 	 * Retrieves document path for given ZoomInto info.
 	 */
-	public DocumentPath getDocumentPath(@NonNull final DocumentZoomIntoInfo zoomIntoInfo)
+	public DocumentPath getDocumentPath(
+			@NonNull final DocumentZoomIntoInfo zoomIntoInfo)
 	{
 		if (!zoomIntoInfo.isRecordIdPresent())
 		{
@@ -650,15 +665,16 @@ public class DocumentCollection
 		}
 
 		//
-		// Invalidate the root documents
+			// Invalidate the root documents
 		rootDocuments.invalidateAll(documentKeys);
 
 		//
-		// Notify frontend
+			// Notify frontend
 		documentKeys.forEach(documentKey -> websocketPublisher.staleRootDocument(documentKey.getWindowId(), documentKey.getDocumentId()));
 	}
 
-	public void invalidateDocumentsByWindowId(@NonNull final WindowId windowId)
+	public void invalidateDocumentsByWindowId(
+			@NonNull final WindowId windowId)
 
 	{
 		final ImmutableList<DocumentKey> documentKeys = rootDocuments.asMap().keySet()
@@ -674,10 +690,13 @@ public class DocumentCollection
 
 	public void invalidateAll(final Collection<DocumentToInvalidate> documentToInvalidateList)
 	{
-		documentToInvalidateList.forEach(this::invalidate);
+		for (final DocumentToInvalidate documentToInvalidate : documentToInvalidateList)
+		{
+			invalidate(documentToInvalidate);
+		}
 	}
 
-	private void invalidate(final DocumentToInvalidate documentToInvalidate)
+	private void invalidate(@NonNull final DocumentToInvalidate documentToInvalidate)
 	{
 		final ImmutableList<DocumentEntityDescriptor> entityDescriptors = getCachedWindowIdsForTableName(documentToInvalidate.getTableName())
 				.stream()
@@ -695,34 +714,34 @@ public class DocumentCollection
 			final WindowId windowId = entityDescriptor.getWindowId();
 			final DocumentKey rootDocumentKey = DocumentKey.of(windowId, rootDocumentId);
 			final Document rootDocument = rootDocuments.getIfPresent(rootDocumentKey);
-			if (rootDocument != null)
-			{
-				try (final IAutoCloseable ignored = rootDocument.lockForWriting())
+				if (rootDocument != null)
 				{
-					for (final IncludedDocumentToInvalidate includedDocumentToInvalidate : documentToInvalidate.getIncludedDocuments())
+					try (final IAutoCloseable ignored = rootDocument.lockForWriting())
 					{
-						final DocumentIdsSelection includedRowIds = includedDocumentToInvalidate.toDocumentIdsSelection();
-						if (includedRowIds.isEmpty())
+						for (final IncludedDocumentToInvalidate includedDocumentToInvalidate : documentToInvalidate.getIncludedDocuments())
 						{
-							continue;
-						}
+							final DocumentIdsSelection includedRowIds = includedDocumentToInvalidate.toDocumentIdsSelection();
+							if (includedRowIds.isEmpty())
+							{
+								continue;
+							}
 
-						for (final DocumentEntityDescriptor includedEntityDescriptor : entityDescriptor.getIncludedEntitiesByTableName(includedDocumentToInvalidate.getTableName()))
-						{
-							final DetailId detailId = includedEntityDescriptor.getDetailId();
+							for (final DocumentEntityDescriptor includedEntityDescriptor : entityDescriptor.getIncludedEntitiesByTableName(includedDocumentToInvalidate.getTableName()))
+							{
+								final DetailId detailId = includedEntityDescriptor.getDetailId();
 
-							rootDocument.getIncludedDocumentsCollection(Check.assumeNotNull(detailId, "Expected detailId not null")).markStale(includedRowIds);
+								rootDocument.getIncludedDocumentsCollection(Check.assumeNotNull(detailId, "Expected detailId not null")).markStale(includedRowIds);
+							}
 						}
 					}
 				}
-			}
 
-			//
-			// Invalidate the root document
-			if (documentToInvalidate.isInvalidateDocument())
-			{
+				//
+				// Invalidate the root document
+				if (documentToInvalidate.isInvalidateDocument())
+				{
 				rootDocuments.invalidate(rootDocumentKey);
-			}
+				}
 
 			//
 			// Notify frontend, even if the root document does not exist (or it was not cached).
@@ -758,7 +777,8 @@ public class DocumentCollection
 	/**
 	 * Invalidates all root documents identified by tableName/recordId and notifies frontend (via websocket).
 	 */
-	public void invalidateRootDocument(@NonNull final DocumentPath documentPath)
+	public void invalidateRootDocument(
+			@NonNull final DocumentPath documentPath)
 	{
 		final DocumentKey documentKey = DocumentKey.ofRootDocumentPath(documentPath);
 
@@ -802,7 +822,10 @@ public class DocumentCollection
 		return DocumentPath.rootDocumentPath(fromDocumentPath.getWindowId(), DocumentId.of(toPO.get_ID()));
 	}
 
-	public void duplicateTabRowInTrx(@NonNull final TableRecordReference parentRef, @NonNull final TableRecordReference fromRecordRef, @NonNull final AdWindowId windowId)
+	public void duplicateTabRowInTrx(
+			@NonNull final TableRecordReference parentRef,
+			@NonNull final TableRecordReference fromRecordRef,
+			@NonNull final AdWindowId windowId)
 	{
 		final Object fromModel = fromRecordRef.getModel(PlainContextAware.newWithThreadInheritedTrx());
 		final String tableName = InterfaceWrapperHelper.getModelTableName(fromModel);
@@ -872,13 +895,13 @@ public class DocumentCollection
 	@Immutable
 	private static final class DocumentKey
 	{
-		public static DocumentKey of(final Document document)
+		public static DocumentKey of(@NonNull final Document document)
 		{
 			final DocumentPath documentPath = document.getDocumentPath();
 			return ofRootDocumentPath(documentPath);
 		}
 
-		public static DocumentKey ofRootDocumentPath(final DocumentPath documentPath)
+		public static DocumentKey ofRootDocumentPath(@NonNull final DocumentPath documentPath)
 		{
 			if (!documentPath.isRootDocument())
 			{
@@ -888,10 +911,16 @@ public class DocumentCollection
 			{
 				throw new InvalidDocumentPathException(documentPath, "document path for creating new documents is not allowed");
 			}
-			return new DocumentKey(documentPath.getDocumentType(), documentPath.getDocumentTypeId(), documentPath.getDocumentId());
+
+			return new DocumentKey(
+					documentPath.getDocumentType(),
+					documentPath.getDocumentTypeId(),
+					documentPath.getDocumentId());
 		}
 
-		public static DocumentKey of(@NonNull final WindowId windowId, @NonNull final DocumentId documentId)
+		public static DocumentKey of(
+				@NonNull final WindowId windowId,
+				@NonNull final DocumentId documentId)
 		{
 			return new DocumentKey(DocumentType.Window, windowId.toDocumentId(), documentId);
 		}
@@ -902,11 +931,14 @@ public class DocumentCollection
 
 		private Integer _hashcode = null;
 
-		private DocumentKey(final DocumentType documentType, final DocumentId documentTypeId, final DocumentId documentId)
+		private DocumentKey(
+				@NonNull final DocumentType documentType,
+				@NonNull final DocumentId documentTypeId,
+				@NonNull final DocumentId documentId)
 		{
-			this.documentType = Preconditions.checkNotNull(documentType, "documentType");
-			this.documentTypeId = Preconditions.checkNotNull(documentTypeId, "documentTypeId");
-			this.documentId = Preconditions.checkNotNull(documentId, "documentId");
+			this.documentType = documentType;
+			this.documentTypeId = documentTypeId;
+			this.documentId = documentId;
 		}
 
 		@Override
